java et script

+++ WEB-INF/classes/FleursCategories.java
@@ -2,7 +2,7 @@ package fleur;
 
 public class FleursCategories {
     private String nomFleur;
-    private double prix;
+    private int prix;
     private String descriptions;
     private String images;
     private int categoryId;
@@ -12,7 +12,7 @@ public class FleursCategories {
 
     public FleursCategories() {
     }
-    public FleursCategories(String nomFleur, double prix, String descriptions, String images, int categoryId,
+    public FleursCategories(String nomFleur, int prix, String descriptions, String images, int categoryId,
             String nomCategory) {
         this.nomFleur = nomFleur;
         this.prix = prix;
